+commit 03.03.24

+++ Books/App.cs
@@ -14,12 +14,27 @@ namespace Books
 
     public class App
     {
-        private Filter _appFilter;
-        private LibraryRepository _appLibraryRepository;
-        private FileReader _fileReader;
+        private readonly Filter _appFilter;
+        private readonly LibraryRepository _appLibraryRepository;
+        private readonly FileReader _fileReader;
 
         public App(Filter filter, LibraryRepository libraryRepository, FileReader fileReader) 
         {
+            if(filter == null)
+            {
+                throw new ArgumentNullException(nameof(filter), "Filter is null");
+            }
+
+            if(libraryRepository == null)
+            {
+                throw new ArgumentNullException(nameof(libraryRepository), "Library repository is null");
+            }
+
+            if(fileReader == null)
+            {
+                throw new ArgumentNullException(nameof(fileReader), "File reader is null");
+            }
+
             _appFilter = filter;
             _appLibraryRepository = libraryRepository;
             _fileReader = fileReader;
@@ -27,8 +42,45 @@ namespace Books
 
         public void Run(string[] args)
         {
+            if (args == null)
+            {
+                throw new ArgumentNullException(nameof(args), "Args is null");
+            }
+
             _appLibraryRepository.DeleteDB();
             _appLibraryRepository.CreateDB();
+
+            string filePath = GetPath(args);
+
+            BookModel[] bookModels = _fileReader.Read(filePath).ToArray();
+
+            List<GenreEntity> genreEntities = GenreMapper.GetEntities(bookModels).ToList();
+            List<AuthorEntity> authorEntities = AuthorMapper.GetEntities(bookModels).ToList();
+            List<PublisherEntity> publisherEntities = PublisherMapper.GetEntities(bookModels).ToList();
+
+            List<BookEntity> bookEntities = BookMapper.GetEntities(bookModels, genreEntities, authorEntities, publisherEntities).ToList();
+
+            foreach (var bookEntity in bookEntities)
+            {
+                _appLibraryRepository.AddBook(bookEntity);
+            }
+
+            IReadOnlyList<BookEntity> filteredBooksEntities = _appLibraryRepository.FindBooks(_appFilter);
+
+            List<BookModel> result = BookMapper.GetModels(filteredBooksEntities);
+
+            string nameOfDirectory = PathIO.GetDirectoryName(filePath);
+
+            Printer.PrintResultInConsole(result);
+            Printer.PrintResultsToFile(nameOfDirectory, result);
+        }
+
+        private string GetPath(string[] args)
+        {
+            if (args == null)
+            {
+                throw new ArgumentNullException(nameof(args), "Args is null");
+            }
 
             string filePath = Path.GetFromArgs(args, nameof(filePath));
 
@@ -39,27 +91,7 @@ namespace Books
                 filePath = Path.GetFromUserInput();
             }
 
-            BookModel[] bookModels = _fileReader.Read(filePath).ToArray();
-
-            List<GenreEntity> genreEntities = BookModelsToGenreEntities.Convert(bookModels);
-            List<AuthorEntity> authorEntities = BookModelsToAuthorEntities.Convert(bookModels);
-            List<PublisherEntity> publisherEntities = BookModelsToPublisherEntities.Convert(bookModels);
-
-            List<BookEntity> bookEntities = BookModelsToBookEntities.Convert(bookModels, genreEntities, authorEntities, publisherEntities);
-
-            foreach (var bookEntity in bookEntities)
-            {
-                _appLibraryRepository.AddBook(bookEntity);
-            }
-
-            IReadOnlyList<BookEntity> filteredBooksEntities = _appLibraryRepository.FindBooks(_appFilter);
-
-            List<BookModel> result = BookEntitiesToBookModels.Convert(filteredBooksEntities);
-
-            string nameOfDirectory = PathIO.GetDirectoryName(filePath);
-
-            Printer.PrintResultInConsole(result);
-            Printer.PrintResultsToFile(nameOfDirectory, result);
+            return filePath;
         }
     }
 }
